requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.1
APScheduler==3.10.4
playwright==1.55.0
Jinja2==3.1.4
python-dateutil==2.9.0.post0
tzdata==2025.2
